TextureLoader.cs を class 別に分解
