+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/TextureLoader.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54a405d31cce94643be8f76f805da8c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
